requirements.txt
fastapi~=0.104.0
Markdown~=3.5
python-dotenv~=1.0.0
starlette~=0.27.0